feat : #99 - complete update logic before using comment

+++ backend/tumblbug-clone/src/main/java/com/example/tumblbugclone/service/ProjectUpdateService.java
@@ -1,18 +1,21 @@
 package com.example.tumblbugclone.service;
 
 import com.example.tumblbugclone.Exception.updateException.CantFindUpdateException;
-import com.example.tumblbugclone.Exception.updateException.UpdateCantModifyIdException;
 import com.example.tumblbugclone.Exception.updateException.UpdateCantModifyModifiedToFalse;
+import com.example.tumblbugclone.Exception.userexception.UnauthorizedUserException;
+import com.example.tumblbugclone.Exception.userexception.UserCantFindException;
 import com.example.tumblbugclone.dto.ProjectUpdateDTO;
+import com.example.tumblbugclone.dto.UserSendingDTO;
 import com.example.tumblbugclone.model.Project;
 import com.example.tumblbugclone.model.ProjectUpdate;
 import com.example.tumblbugclone.model.User;
 import com.example.tumblbugclone.repository.ProjectRepository;
 import com.example.tumblbugclone.repository.ProjectUpdateRepository;
-import com.example.tumblbugclone.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -21,18 +24,37 @@ public class ProjectUpdateService {
 
     private final ProjectUpdateRepository projectUpdateRepository;
     private final ProjectRepository projectRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
 
     @Autowired
-    public ProjectUpdateService(ProjectUpdateRepository projectUpdateRepository, ProjectRepository projectRepository, UserRepository userRepository){
+    public ProjectUpdateService(ProjectUpdateRepository projectUpdateRepository, ProjectRepository projectRepository, UserService userService){
         this.projectUpdateRepository = projectUpdateRepository;
         this.projectRepository = projectRepository;
-        this.userRepository = userRepository;
+        this.userService = userService;
     }
 
-    public long save(long userIndex, long projectId, ProjectUpdateDTO projectUpdateDTO){
-        ProjectUpdate projectUpdate = convertDTO2Update(userIndex, projectId, projectUpdateDTO);
+    public long save(long userIndex, long projectId, String content) throws UserCantFindException {
+        User user;
+        ProjectUpdateDTO projectUpdateDTO = new ProjectUpdateDTO();
+        try{
+            user = userService.findUserByIndex(userIndex);
+        }catch (EmptyResultDataAccessException e){
+            throw new UserCantFindException();
+        }
+
+        Project project = null;
+        try{
+            project = projectRepository.findProjectById(projectId);
+        }catch(EmptyResultDataAccessException e){
+            //추후 추가 예정
+            System.out.println("project is not exist");
+        }
+
+        projectUpdateDTO.setContent(content);
+        projectUpdateDTO.setUpdateDate(new Date());
+
+        ProjectUpdate projectUpdate = convertDTO2Update(user, project, projectUpdateDTO);
         long updateId = projectUpdateRepository.save(projectUpdate.getProject(), projectUpdate);
         return updateId;
     }
@@ -44,21 +66,26 @@ public class ProjectUpdateService {
         return convertUpdate2DTO(byId);
     }
 
-    public void update(ProjectUpdateDTO projectUpdateDTO) throws UpdateCantModifyModifiedToFalse {
-        //사용자의 수정 권한 확인 필요
-        ProjectUpdate originalUpdate = projectUpdateRepository.findById(projectUpdateDTO.getId());
-        if(originalUpdate.isModified() && !projectUpdateDTO.isModified())
-            throw new UpdateCantModifyModifiedToFalse();
-        originalUpdate.setUpdateDate(projectUpdateDTO.getUpdateDate());
-        originalUpdate.setContent(projectUpdateDTO.getContent());
+    public ProjectUpdateDTO update(long userIdx, long ContentId, String newContent) throws UnauthorizedUserException {
+        ProjectUpdate originalUpdate = projectUpdateRepository.findById(ContentId);
+        if(originalUpdate.getCreater().getUserIdx() != userIdx)
+            throw new UnauthorizedUserException();
+        originalUpdate.setUpdateDate(new Date());
+        originalUpdate.setContent(newContent);
         originalUpdate.setModified(true);
 
         projectUpdateRepository.update(originalUpdate);
+
+        return convertUpdate2DTO(originalUpdate);
     }
 
-    public void delete(long projectId){
+    public void delete(long userIdx, long ContentId) throws UnauthorizedUserException, CantFindUpdateException {
         //사용자의 삭제 권한 확인 필요
-        ProjectUpdate updateById = projectUpdateRepository.findById(projectId);
+        ProjectUpdate updateById = projectUpdateRepository.findById(ContentId);
+        if(updateById == null)
+            throw new CantFindUpdateException();
+        if(userIdx != updateById.getCreater().getUserIdx())
+            throw new UnauthorizedUserException();
         projectUpdateRepository.remove(updateById);
     }
 
@@ -74,9 +101,8 @@ public class ProjectUpdateService {
 
         return dtos;
     }
-    private ProjectUpdate convertDTO2Update(long userIndex, long projectId, ProjectUpdateDTO projectUpdateDTO){
-        User user = userRepository.findUserByIndex(userIndex);
-        Project project = projectRepository.findProjectById(projectId);
+
+    private ProjectUpdate convertDTO2Update(User user, Project project, ProjectUpdateDTO projectUpdateDTO){
         ProjectUpdate projectUpdate = new ProjectUpdate();
         projectUpdate.setProject(project);
         projectUpdate.setCreater(user);
@@ -84,19 +110,21 @@ public class ProjectUpdateService {
         projectUpdate.setUpdateDate(projectUpdateDTO.getUpdateDate());
         projectUpdate.setContent(projectUpdateDTO.getContent());
 
-
         return projectUpdate;
     }
 
     private ProjectUpdateDTO convertUpdate2DTO(ProjectUpdate projectUpdate) {
+        UserSendingDTO creater = userService.findSendingUserByIndex(projectUpdate.getCreater().getUserIdx());
+
         ProjectUpdateDTO projectUpdateDTO = new ProjectUpdateDTO();
         projectUpdateDTO.setId(projectUpdate.getId());
         projectUpdateDTO.setUpdateDate(projectUpdate.getUpdateDate());
-        projectUpdateDTO.setCreaterName(projectUpdate.getCreater().getUserName());
         projectUpdateDTO.setContent(projectUpdate.getContent());
         projectUpdateDTO.setProjectId(projectUpdate.getProject().getProjectId());
         projectUpdateDTO.setModified(projectUpdate.isModified());
+        projectUpdateDTO.setCreater(creater);
 
         return projectUpdateDTO;
     }
+
 }
